Use a single CentralDogma watcher per file
